requirements.txt
tensorflow
streamlit
pillow
pandas
matplotlib
numpy
requests

#Run this file with - pip3 install -r requirements.txt